Auth Login Successful
Popup works, need to fix login status reactivity
