+++ nuxt.config.ts
@@ -1,4 +1,16 @@
 import { defineNuxtConfig } from "nuxt3";
 
 // https://v3.nuxtjs.org/docs/directory-structure/nuxt.config
-export default defineNuxtConfig({});
+export default defineNuxtConfig({
+  ssr: false,
+  runtimeConfig: {
+    // The private keys which are only available within server-side
+
+    // Keys within public, will be also exposed to the client-side
+    public: {
+      auth0Domain: process.env.AUTH0_DOMAIN,
+      auth0ClientId: process.env.AUTH0_CLIENT_ID,
+      apiBase: "/api",
+    },
+  },
+});
